theater list
some html js css update
