+++ public/assets/css/theater/theater.css
@@ -1,17 +1,21 @@
-.theatre {
+.theater_section {
+    margin-top: 61px;
+}
+
+.theater {
     display: flex;
     flex-direction: column;
     width: 80%;
     margin: auto;
 }
 
-.theatre-calendar-wrapper {
+.theater-calendar-wrapper {
     position: relative;
     overflow: hidden;
     margin-bottom: 10px;
 }
 
-.theatre-calendar-list {
+.theater-calendar-list {
     display: flex;
     flex-direction: row;
     width: max-content;
@@ -20,14 +24,14 @@
     padding: 0;
 }
 
-.theatre-calendar-date {
+.theater-calendar-date {
     width: 62px;
     height: 70px;
     margin-right: 8px;
     vertical-align: top;
 }
 
-.theatre-calendar-date_btn {
+.theater-calendar-date_btn {
     width: 62px;
     height: 70px;
     padding: 0;
@@ -39,13 +43,13 @@
     text-transform: uppercase;
 }
 
-.theatre-calendar-date_select {
+.theater-calendar-date_select {
     border-color: #3478c1;
     background-color: #3478c1;
     color: #fff;
 }
 
-.theatre-calendar-date_btn_weekday {
+.theater-calendar-date_btn_weekday {
     font-family: proxima-nova,Helvetica,Arial,sans-serif;
     font-size: 12px;
     font-weight: 700;
@@ -54,7 +58,7 @@
     height: 12px;
 }
 
-.theatre-calendar-date_btn_month {
+.theater-calendar-date_btn_month {
     font-size: 12px;
     letter-spacing: .3px;
     margin-top: 5px;
@@ -65,7 +69,7 @@
     overflow: hidden;
 }
 
-.theatre-calendar-date_btn_day {
+.theater-calendar-date_btn_day {
     font-size: 25px;
     font-family: "Futura Condensed","Arial Narrow",Arial,sans-serif;
     color: #171c20;
@@ -78,13 +82,13 @@
     color: #fff;
 }
 
-.theatre-detail {
+.theater-detail {
     position: relative;
     display: flex;
     flex-direction: column;
 }
 
-.theatre-detail-showtime {
+.theater-detail-showtime {
     margin-top: 20px;
     margin-bottom: 0;
     padding: 20px;
@@ -94,16 +98,16 @@
     position: relative;
 }
 
-.theatre-detail-showtime_header {
+.theater-detail-showtime_header {
     padding-bottom: 40px;
     position: relative;
 }
 
-.theatre-detail-showtime_header_name {
+.theater-detail-showtime_header_name {
     display: flex;
 }
 
-.theatre-detail-showtime_header_name h2 {
+.theater-detail-showtime_header_name h2 {
     font-size: 1rem;
     font-weight: 700;
     overflow-wrap: break-word;
@@ -111,19 +115,19 @@
     margin: 0;
 }
 
-.theatre-detail-showtime_header_address {
+.theater-detail-showtime_header_address {
     margin-top: 10px;
     display: inline-block;
 }
 
-.theatre-detail-showtime_header_address p {
+.theater-detail-showtime_header_address p {
     font-size: 14px;
     font-weight: 600;
     color: #3478c1;
     margin: 0;
 }
 
-.theatre-detail-showtime_variants_format {
+.theater-detail-showtime_variants_format {
     border-top: 0;
     margin-top: 0;
     margin-bottom: 16px;
@@ -134,14 +138,14 @@
     line-height: 20px;
 }
 
-.theatre-detail-showtime_variants_list_wrapper {
+.theater-detail-showtime_variants_list_wrapper {
     margin-top: 10px;
     margin: 0;
     padding: 0;
     min-width: 300px;
 }
 
-.theatre-detail-showtime_variants_list {
+.theater-detail-showtime_variants_list {
     margin: 0;
     padding: 0;
     list-style-type: none;
@@ -149,7 +153,7 @@
     overflow: hidden;
 }
 
-.theatre-detail-showtime_variants_time {
+.theater-detail-showtime_variants_time {
     display: inline-block;
     margin-bottom: 10px;
     margin-right: 10px;
@@ -157,7 +161,7 @@
     height: 35px;
 }
 
-.theatre-detail-showtime_variants_time_btn{
+.theater-detail-showtime_variants_time_btn{
     width: 78px;
     height: 35px;
     display: flex;
