+++ routes/main.js
@@ -1,6 +1,7 @@
 const staticRoutes = require('./static');
 const userRoutes = require('./user');
 const detailsRoutes = require('./details');
+const theaterRoutes = require('./details');
 
 const initRoutes = function (app) {
     console.log("Initializing Routes...");
@@ -8,6 +9,7 @@ const initRoutes = function (app) {
     app.use('/', staticRoutes); // Static + Home + Auth
     app.use('/users', userRoutes); // Users
     app.use('/details', detailsRoutes); //details
+    app.use('/theater', theaterRoutes)
 
     // Ending Routes
     console.log('Finished Initializing Routes...');
