+++ views/pages/theater/list.handlebars
@@ -1,25 +1,24 @@
-<section>
-    <div class="theatre">
-        <div style="display: none;"><p id="_id">{{{_id}}}</p></div>
-        <section class="theatre-calendar-wrapper">
-            <div class="theatre-calendar">
+<section class="theater_section">
+    <div class="theater">
+        <section class="theater-calendar-wrapper">
+            <div class="theater-calendar">
                 <div style="display: none;"><p id="selected">0</p></div>
                 <div style="display: none;"><p id="selected_day">0</p></div>
                 <div style="display: none;"><p id="selected_month">0</p></div>
                 <div style="display: none;"><p id="selected_year">0</p></div>
-                <ul id="calendar" class="theatre-calendar-list">
+                <ul id="calendar" class="theater-calendar-list">
                     
                 </ul>
             </div>
         </section>
-        <section id="theatre-detail" class="theatre-detail">
+        <section id="theater-detail" class="theater-detail">
         </section>
     </div>
 </section>
 
 
 {{#section "styles"}}
-    <link rel="stylesheet" href="/public/assets/css/theatre/theatre.css">
-    <script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc=" crossorigin="anonymous"></script>
-    <script src="/public/assets/js/theatre/theatre.js"></script>
+    <link rel="stylesheet" href="/public/assets/css/theater/theater.css">
 {{/section}}
+<script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc=" crossorigin="anonymous"></script>
+<script src="/public/assets/js/theater/theater.js"></script>
